up communities

+++ Terradue.Tep/Terradue/Tep/WebServer/Community/Community.Service.cs
@@ -149,7 +149,7 @@ namespace Terradue.Tep.WebServer.Services{
             context.LogInfo (this, string.Format ("/community/search GET"));
 
             EntityList<ThematicCommunity> domains = new EntityList<ThematicCommunity> (context);
-            domains.SetFilter("Kind", DomainKind.Public + "," + DomainKind.Private);
+            domains.SetFilter("Kind", (int)DomainKind.Public + "," + (int)DomainKind.Private);
 
             // Load the complete request
             HttpRequest httpRequest = HttpContext.Current.Request;
